posix: add 6 Linux header constant modules (batch 336)
- linux_bcache_user_types: /sys/fs/bcache sysfs surface — cache-set
  attribute names (register/register_quiet/unregister/flash_vol_create
  /clear_stats/average_key_size/dirty_data/root_usage_percent/
  cache_available_percent), backing-device attrs (attach/detach/stop/
  running/label), cache_mode string values (writethrough/writeback/
  writearound/none), bucket size bounds 16 KiB..16 MiB (powers of two,
  1024x ratio).

- linux_befs_user_types: BeOS filesystem on-disk constants — \"befs\"
  fstype, superblock magic 0x42465331 (\"BFS1\" big-endian ASCII) at
  byte 512 plus two integrity magics, distinct inode + btree magics,
  block size 1 KiB..8 KiB powers of two, inode flag single-bits with
  PERMANENT_FLAG aliasing IN_USE, NAME_LEN=255, SYMLINK_LEN=160.

- linux_binfmt2_user_types: ELF/binfmt loader limits — MAX_ARG_PAGES=32
  (page count, page-size agnostic), MAX_ARG_STRINGS=i32::MAX,
  MAX_ARG_STRLEN=128 KiB, BINPRM_BUF_SIZE=256, MAX_INTERP_RECURSION=5,
  ELF e_type dense 0..4 (NONE/REL/EXEC/DYN/CORE), EXSTACK_* codes with
  DEFAULT==STACK_TOP_TASK_SIZE==0, stack-randomisation masks (22 bits
  on 64-bit, 11 bits on 32-bit, contiguous low-bit runs).

- linux_binfmt_user_types: /proc/sys/fs/binfmt_misc registration —
  register/status control files, ':' default delimiter,
  FIELD_MAX=4096, rule type tags 'E'/'M' (extension/magic), flag
  letters P/O/C/F (preserve-argv0/open/credentials/fix-binary),
  entry-control byte strings \"1\"/\"0\"/\"-1\" (enable/disable/remove).

- linux_bio_user_types: block-layer bi_opf/bi_flags — operation codes
  in low byte (READ=0, WRITE=1, zone family clustered 10..17 with
  OPEN/CLOSE/FINISH/APPEND dense 10..13), OP_MASK=0xFF, request flag
  single-bits in bits 8..24, FAILFAST_MASK unions three bits,
  NOMERGE_FLAGS blocks three bits (NOMERGE/PREFLUSH/FUA).

- linux_blk_cgroup_user_types: cgroup v2 io.* control surface —
  control file names (weight/max/stat/pressure/latency/cost.qos/
  cost.model/bfq.weight), weight 1..10000 with default 100 (geometric
  midpoint), io.max keys paired r/w (rbps/wbps, riops/wiops) with
  \"max\" sentinel, io.stat keys split into bytes-suffixed and
  ios-suffixed groups, ':' device separator (8:0 major:minor form).

All six register in posix/src/lib.rs in strict alphabetical order
(digit < '_' < letter, '_t' types < '_u' user_types). All per-module
tests pass on x86_64-pc-windows-gnu; bare-metal build clean on
x86_64-unknown-none.

+++ gui/desktop/src/power.rs
@@ -889,7 +889,7 @@ impl ScreenSaver {
         self.columns.clear();
         let col_count = self.width / 14; // ~14px per character column
         for i in 0..col_count {
-            let len = (self.next_random_u32() % 20 + 5) as u32;
+            let len = (self.next_random_u32() % 20 + 5);
             let mut chars = Vec::with_capacity(len as usize);
             for _ in 0..len {
                 chars.push((self.next_random_u32() % 94 + 33) as u8); // printable ASCII
